add ExtraCoin
This PR is an application of two Paseo validator nodes maintained by ExtraCoin to be selected to join the active set on Paseo.

ExtraCoin is running invulnerable collators in Kusama Bridge Hub and Polkadot Collectives as per the Paseo selection criteria. 

Following Paradox | ParaNodes.io's call for applications for Paseo Validators in the System Parachain Collators matrix channel.